Separates typing modules
